(#602) code 修复

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/topdynamic/DynamicTopFragment.java
@@ -107,7 +107,7 @@ public class DynamicTopFragment extends TSFragment<DynamicTopContract.Presenter>
     }
 
     @Override
-    public float getInputMoney() {
+    public double getInputMoney() {
         return mInputMoney;
     }
